[chore][feat]: Update file migrate v2_v3 update config redis all api and Add CURD book

+++ src/main/resources/db/migration/V2__create_book_table.sql
@@ -1,5 +1,5 @@
 create table books (
-                       id numeric,
+                       id SERIAL ,
                        title varchar(128),
                        author varchar(256),
                        constraint pk_book primary key (id)
